refactor: reducer 모듈 분리
